examples/text_to_image/requirements.txt
accelerate>=0.16.0
torchvision
transformers>=4.25.1
datasets
ftfy
tensorboard
Jinja2
torchmetrics==0.11.4
gradio
Pillow==9.5.0
